requirements.txt
ultralytics
pandas
numpy
opencv-python
torch
scikit-learn
jupyterlab
streamlit
pillow
matplotlib